pyproject.toml
[tool.poetry]
name = "codanim"
version = "0.1.0"
description = "A framework for creating programmatic animations."
authors = ["Pedro Batista <pedrovhb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyglet = "^1.5.18"
jurigged = "^0.3.4"
attrs = "^21.2.0"
snoop = "^0.4.0"
palanteer = "^0.2.0"
textual = "^0.1.9"
networkx = "^2.6.2"
numpy = "^1.21.1"
ipython = "^7.26.0"
ipykernel = "^6.0.3"
madbg = "^1.1.4"
pdbpp = "^0.10.3"
Rx = "^3.2.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
